Reorganize homepage refresh components
