+++ src/applications/vaos/appointment-list/components/AppointmentsPageV2/index.jsx
@@ -23,7 +23,7 @@ import DowntimeNotification, {
 } from 'platform/monitoring/DowntimeNotification';
 import WarningNotification from '../../../components/WarningNotification';
 import Select from '../../../components/Select';
-import ScheduleNewAppointmentRadioButtons from './ScheduleNewAppointmentRadioButtons';
+import ScheduleNewAppointmentRadioButtons from '../ScheduleNewAppointmentRadioButtons';
 
 const pageTitle = 'VA appointments';
